requirements.txt
feedparser
trafilatura
pyyaml
